refactor: extracted playlist reducer
