+++ src/contexts/playlistContext.js
@@ -1,27 +1,12 @@
 import { useReducer, useEffect, createContext, useContext } from "react";
 import { getAllPlaylists } from "../services/playlistServices";
+import { playlistReducer } from "../reducers/playlistReducer";
+import { UPDATE_PLAYLISTS } from "../reducers/constant";
 import { useAuth } from "./authContext";
 
 const PlaylistContext = createContext(null);
 
 const usePlaylist = () => useContext(PlaylistContext);
-
-const playlistReducer = (state, action) => {
-  switch (action.type) {
-    case "UPDATE_PLAYLISTS":
-      return { ...state, playlists: [...action.payload] };
-
-    case "UPDATE_SINGLE_PLAYLIST":
-      const { _id: playlistID } = action.payload;
-      const updatedPlaylists = state.playlists.map(playlist =>
-        playlist._id === playlistID ? action.payload : playlist
-      );
-      return { ...state, playlists: updatedPlaylists };
-
-    default:
-      return state;
-  }
-};
 
 const PlaylistProvider = ({ children }) => {
   const {
@@ -38,14 +23,14 @@ const PlaylistProvider = ({ children }) => {
           const response = await getAllPlaylists(token);
           if (response.status === 200) {
             const playlists = response.data.playlists;
-            playlistDspatch({ type: "UPDATE_PLAYLISTS", payload: playlists });
+            playlistDspatch({ type: UPDATE_PLAYLISTS, payload: playlists });
           }
         } catch (error) {
           console.error("ERROR", error);
         }
       })();
     } else {
-      playlistDspatch({ type: "UPDATE_PLAYLISTS", payload: [] });
+      playlistDspatch({ type: UPDATE_PLAYLISTS, payload: [] });
     }
   }, [isAuthenticated, token]);
