+++ src/helpers/playlistHelper/addVideoToPlaylist.js
@@ -1,4 +1,5 @@
 import { addPlaylistVideo } from "../../services/playlistServices";
+import { UPDATE_SINGLE_PLAYLIST } from "../../reducers/constant";
 
 export const addVideoToPlaylist = async (
   token,
@@ -10,7 +11,7 @@ export const addVideoToPlaylist = async (
     const response = await addPlaylistVideo(token, video, playlist);
     if (response.status === 201) {
       playlistDspatch({
-        type: "UPDATE_SINGLE_PLAYLIST",
+        type: UPDATE_SINGLE_PLAYLIST,
         payload: response.data.playlist,
       });
     }
