+++ src/reducers/constant.js
@@ -1,0 +1,3 @@
+// Playlist Reducer
+export const UPDATE_PLAYLISTS = "UPDATE_PLAYLISTS";
+export const UPDATE_SINGLE_PLAYLIST = "UPDATE_SINGLE_PLAYLIST";
